show initiatives from database on the map

+++ db/migrate/20191005213108_add_location_coords_to_initiatives.rb
@@ -2,7 +2,7 @@
 
 class AddLocationCoordsToInitiatives < ActiveRecord::Migration[6.0]
   def change
-    add_column :initiatives, :latitude, :integer
-    add_column :initiatives, :longitude, :integer
+    add_column :initiatives, :latitude, :float
+    add_column :initiatives, :longitude, :float
   end
 end
